% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_agresti_coull_ci}
\alias{calc_agresti_coull_ci}
\title{Calculate Agresti-Coull Confidence Interval}
\usage{
calc_agresti_coull_ci(p_hat, n, conf_level = 0.95)
}
\arguments{
\item{p_hat}{\link{numeric} Sample proportion}

\item{n}{\link{integer} Sample size}

\item{conf_level}{\link{numeric} Confidence level (default 0.95)}
}
\value{
\link{numeric vector} Lower and upper bounds
}
\description{
Calculate Agresti-Coull Confidence Interval
}
