% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_sample_house_edge}
\alias{calc_sample_house_edge}
\title{Calculate the actual sample house edge}
\usage{
calc_sample_house_edge(sim_data, wager_per_round)
}
\arguments{
\item{sim_data}{data.frame Simulation data with 'NetWin' column}

\item{wager_per_round}{numeric The amount bet per round}
}
\value{
numeric The sample house edge
}
\description{
Calculate the actual sample house edge
}
