% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_sample_probability}
\alias{calc_sample_probability}
\title{Calculate sample probability for a target match count}
\usage{
calc_sample_probability(sim_data, target_matches)
}
\arguments{
\item{sim_data}{\link{data.frame} Simulation data with a 'Matches' column}

\item{target_matches}{\link{integer} The match count to calculate probability for}
}
\value{
\link{numeric} The sample probability
}
\description{
Calculate sample probability for a target match count
}
