% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_win_rate}
\alias{calc_win_rate}
\title{Calculate the actual sample win rate}
\usage{
calc_win_rate(sim_data)
}
\arguments{
\item{sim_data}{[data.frame] Simulation data with 'NetWin' column}
}
\value{
[numeric] The sample win rate
}
\description{
Calculate the actual sample win rate
}
