% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{format_ci_vector}
\alias{format_ci_vector}
\title{Format a confidence interval for display from a vector}
\usage{
format_ci_vector(ci)
}
\arguments{
\item{ci}{[numeric vector] Lower and upper bounds}
}
\value{
[character] Formatted interval string
}
\description{
Format a confidence interval for display from a vector
}
