% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{format_confidence_interval}
\alias{format_confidence_interval}
\title{Format a confidence interval for display}
\usage{
format_confidence_interval(p_hat, moe)
}
\arguments{
\item{p_hat}{numeric Sample proportion}

\item{moe}{numeric Margin of Error}
}
\value{
character Formatted interval string
}
\description{
Format a confidence interval for display
}
