% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_sample_distribution}
\alias{get_sample_distribution}
\title{Calculate the distribution of outcomes from simulation data}
\usage{
get_sample_distribution(sim_data)
}
\arguments{
\item{sim_data}{data.frame Simulation data with 'Matches' column}
}
\value{
data.frame Relative frequency distribution
}
\description{
Calculate the distribution of outcomes from simulation data
}
