% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_theoretical_win_rate}
\alias{get_theoretical_win_rate}
\title{Get the theoretical win rate}
\usage{
get_theoretical_win_rate(payouts = DEFAULT_PAYOUT_MULTIPLIERS)
}
\arguments{
\item{payouts}{numeric vector Named vector of payout multipliers}
}
\value{
numeric The theoretical win rate
}
\description{
The probability of rolling at least one match \eqn{X \ge 1}:
\deqn{P(X \geq 1) = \sum_{x=1}^{3} P(X=x)}
}
