% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ci_behavior}
\alias{plot_ci_behavior}
\title{Plot CI Behavior as sample size grows}
\usage{
plot_ci_behavior(sim_data, target_matches)
}
\arguments{
\item{sim_data}{\link{data.frame} Simulation data}

\item{target_matches}{\link{integer} The match count}
}
\value{
\link[ggplot2:ggplot]{ggplot2::ggplot} Ribbon plot
}
\description{
Plot CI Behavior as sample size grows
}
