% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ci_comparison}
\alias{plot_ci_comparison}
\title{Plot Comparison of Confidence Intervals}
\usage{
plot_ci_comparison(p_hat, n, conf_level = 0.95)
}
\arguments{
\item{p_hat}{[numeric] Sample proportion}

\item{n}{[integer] Sample size}

\item{conf_level}{[numeric] Confidence level}
}
\value{
[ggplot] Error bar plot
}
\description{
Plot Comparison of Confidence Intervals
}
