% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_house_edge_convergence}
\alias{plot_house_edge_convergence}
\title{Plot House Edge Convergence}
\usage{
plot_house_edge_convergence(sim_data, theo_edge)
}
\arguments{
\item{sim_data}{data.frame Simulation data}

\item{theo_edge}{numeric Theoretical house edge}
}
\value{
ggplot Convergence plot
}
\description{
Plot House Edge Convergence
}
