% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lln_convergence}
\alias{plot_lln_convergence}
\title{Plot Law of Large Numbers convergence}
\usage{
plot_lln_convergence(sim_data, target_matches)
}
\arguments{
\item{sim_data}{[data.frame] Simulation data}

\item{target_matches}{[integer] The match count to visualize}
}
\value{
[ggplot] Convergence plot
}
\description{
Plot Law of Large Numbers convergence
}
