% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_outcome_distribution}
\alias{plot_outcome_distribution}
\title{Plot the outcome frequency distribution}
\usage{
plot_outcome_distribution(sim_data)
}
\arguments{
\item{sim_data}{\link{data.frame} Simulation data}
}
\value{
\link[ggplot2:ggplot]{ggplot2::ggplot} Distribution bar plot
}
\description{
Plot the outcome frequency distribution
}
