% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_moe_proportion}
\alias{calc_moe_proportion}
\title{Calculate Margin of Error for a proportion}
\usage{
calc_moe_proportion(p_hat, n, conf_level = 0.95)
}
\arguments{
\item{p_hat}{[numeric] Sample proportion}

\item{n}{[integer] Sample size}

\item{conf_level}{[numeric] Confidence level (e.g., 0.95)}
}
\value{
[numeric] The margin of error
}
\description{
Calculate Margin of Error for a proportion
}
