% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_theoretical_house_edge}
\alias{get_theoretical_house_edge}
\title{Get the theoretical house edge}
\usage{
get_theoretical_house_edge(payouts = DEFAULT_PAYOUT_MULTIPLIERS)
}
\arguments{
\item{payouts}{[numeric vector] Named vector of payout multipliers}
}
\value{
[numeric] The theoretical house edge
}
\description{
Get the theoretical house edge
}
