% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_theoretical_prob}
\alias{get_theoretical_prob}
\title{Get theoretical probability for a match count}
\usage{
get_theoretical_prob(matches)
}
\arguments{
\item{matches}{\link{integer} The target number of matches (0-3)}
}
\value{
\link{numeric} The theoretical probability
}
\description{
Get theoretical probability for a match count
}
