% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_theoretical_win_rate}
\alias{get_theoretical_win_rate}
\title{Get the theoretical win rate}
\usage{
get_theoretical_win_rate(payouts = DEFAULT_PAYOUT_MULTIPLIERS)
}
\arguments{
\item{payouts}{\link{numeric vector} Named vector of payout multipliers}
}
\value{
\link{numeric} The theoretical win rate
}
\description{
Get the theoretical win rate
}
