% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lln_convergence}
\alias{plot_lln_convergence}
\title{Plot Law of Large Numbers convergence}
\usage{
plot_lln_convergence(sim_data, target_matches)
}
\arguments{
\item{sim_data}{\link{data.frame} Simulation data}

\item{target_matches}{\link{integer} The match count to visualize}
}
\value{
\link[ggplot2:ggplot]{ggplot2::ggplot} Convergence plot
}
\description{
Plot Law of Large Numbers convergence
}
